Feat: add `formatTokenValue` helper to `utils`
