+++ src/utils/numbers/index.ts
@@ -7,3 +7,4 @@ export const bnLessThan = (a: BigNumberish, b: BigNumberish) =>
   bigNumberify(a).lt(bigNumberify(b));
 
 export const halfPlusOne = (count: number) => Math.floor(count / 2) + 1;
+export { formatTokenValue } from './formatTokenValue';
